backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
selenium>=4.18.0
PyJWT>=2.9.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
requests>=2.32.0
